backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
whitenoise>=6.6
gunicorn>=21.0
